Fix issue when X-RateLimit-Remaining is x-ratelimit-remaining
Something may have changed in CURL or on the Teamwork end because checking for $headers['X-RateLimit-Remaining'] now throw an Undefined index exception. It appears that Header keys are now in lower case.

This PR is fixing the issue by lowering all the header keys before checking for the 'x-ratelimit-remaining'. As it is stated in the HTTP spec that headers are case-insensitive so it is not certain to have it in any specifieed case I guess.